# delivery-microservice/src/main/resources/openAPI.yaml
openapi: 3.0.3
info:
  title: Delivery Microservice
  description: OpenAPI specifications for delivery microservice
  version: 1.0.0
servers:
  - url: 'https://localhost:8082'
...
paths:
...
  /restaurant/{restaurantID}:
    get:
      operationId: getRest
      tags:
        - Vendor
      summary: Get a restaurant object
      description: Fetch a restaurant object.
      parameters:
        - name: restaurantID
          in: path
          description: ID of the restaurant to query.
          required: true
          schema:
            type: string
            format: uuid
        - name: role
          in: query
          description: The role of the user
          required: true
          schema:
            type: string
            enum:
              - vendor
              - courier
              - customer
              - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
              example:
                "200 OK"
        '403':
          description: Forbidden
        '404':
          description: Restaurant not found
...
  /restaurant/{restaurantID}/add/courier/{courierID}:
    patch:
      operationId: addCourierToRest
      tags:
        - Vendor
      summary: Add a courier to a restaurant
      description: A vendor can add a courier to its restaurant
      parameters:
        - in: path
          name: courierID
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the courier to query
        - in: path
          name: restaurantID
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the restaurant to query
        - name: role
          in: query
          description: The role of the user
          required: true
          schema:
            type: string
            enum:
              - vendor
              - courier
              - customer
              - admin
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
              example:
                "error 403"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: string
              example:
                "error 404"
  /restaurant/{restaurantID}/remove/courier/{courierID}:
    patch:
      operationId: removeCourierRest
      tags:
        - Vendor
      summary: Remove a courier from a restaurant
      description: A vendor can remove a courier from its restaurant
      parameters:
        - in: path
          name: courierID
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the courier to query
        - in: path
          name: restaurantID
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the restaurant to query
        - name: role
          in: query
          description: The role of the user
          required: true
          schema:
            type: string
            enum:
              - vendor
              - courier
              - customer
              - admin
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
              example:
                "error 403"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: string
              example:
                "error 404"
components:
...
  securitySchemes:
    UserIdAuth:
      type: apiKey
      in: header
      name: X-User-Id
security:
  - UserIdAuth: [ ]